requirements.txt
PyMuPDF==1.23.21
requests==2.31.0
openai==1.12.0
python-dotenv==1.0.0
tenacity==8.2.2
dash==2.15.0
pandas==2.2
numpy==1.26.4
scikit-learn==1.4
plotly==5.19.0
